Migrate stories for EventPreview to Doctocat #507
