+++ docs/content/components/avatars.mdx
@@ -1,0 +1,7 @@
+---
+title: Avatars
+status: Experimental
+source: https://github.com/UNDataForum/design-system/blob/master/packages/components/src/avatars.js
+---
+
+Placeholder page.
