+++ docs/content/components/email-icon.mdx
@@ -4,7 +4,11 @@ status: Experimental
 source: https://github.com/UNDataForum/design-system/blob/master/packages/components/src/email-icon.js
 ---
 
-An `EmailIcon` is a square inline element. The default size is 32 pixels:
+An `EmailIcon` is a square inline element.
+
+## Examples
+
+The default size is 32 pixels:
 
 ```jsx live
 <EmailIcon />
