+++ docs/content/components/names.mdx
@@ -1,0 +1,7 @@
+---
+title: Names
+status: Experimental
+source: https://github.com/UNDataForum/design-system/blob/master/packages/components/src/names.js
+---
+
+Placeholder page.
