+++ docs/content/components/twitter-icon.mdx
@@ -4,9 +4,11 @@ status: Experimental
 source: https://github.com/UNDataForum/design-system/blob/master/packages/components/src/twitter-icon.js
 ---
 
+A `TwitterIcon` is a square inline element.
+
 ## Examples
 
-A `TwitterIcon` is a square inline element. The default size is 32 pixels:
+The default size is 32 pixels:
 
 ```jsx live
 <TwitterIcon />
