+++ docs/package.json
@@ -21,6 +21,7 @@
     "@undataforum/preset": "*",
     "dotenv": "8.2.0",
     "gatsby": "2.22.12",
+    "gatsby-plugin-theme-ui": "0.3.0",
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
