+++ docs/src/gatsby-plugin-theme-ui/index.js
@@ -1,0 +1,3 @@
+import preset from '@undataforum/preset';
+
+export default preset;
