+++ packages/assets/src/logo.js
